fix(magnet) : fix create/edition modal filled by other magnets
